+++ examples/simple.rs
@@ -18,7 +18,7 @@ extern crate pushrod;
 use piston_window::*;
 use pushrod::core::main::*;
 use pushrod::core::point::*;
-use pushrod::core::window::*;
+use pushrod::core::widget_store::*;
 use pushrod::event::event::*;
 use pushrod::widget::box_widget::*;
 use pushrod::widget::timer_widget::*;
@@ -66,60 +66,53 @@ impl EventListener for ExampleListener {
 }
 
 fn main() {
-    let opengl = OpenGL::V3_2;
-    let prod: Pushrod = Pushrod::new();
-    let mut pushrod_window: PushrodWindow = PushrodWindow::new(
-        WindowSettings::new("Pushrod Window", [800, 600])
-            .opengl(opengl)
-            .resizable(false)
-            .build()
-            .unwrap_or_else(|error| panic!("Failed to build PistonWindow: {}", error)),
-    );
-
-    pushrod_window.window.set_max_fps(60);
-    pushrod_window.window.set_ups(60);
+    let mut prod: Pushrod = Pushrod::new(WindowSettings::new("Pushrod Window", [800, 600])
+        .opengl(OpenGL::V3_2)
+        .resizable(false)
+        .build()
+        .unwrap_or_else(|error| panic!("Failed to build PistonWindow: {}", error)));
 
     let mut base_widget = BaseWidget::new();
     base_widget.set_origin(50, 50);
     base_widget.set_size(200, 200);
     base_widget.set_color([0.5, 0.5, 0.5, 1.0]);
-    pushrod_window.add_widget(Box::new(base_widget));
+    prod.widget_store.add_widget(Box::new(base_widget));
 
     let mut box_widget = BoxWidget::new();
     box_widget.set_origin(275, 50);
     box_widget.set_size(200, 200);
     box_widget.set_color([0.0, 1.0, 0.0, 1.0]);
     box_widget.set_border([1.0, 0.0, 0.0, 1.0], 4);
-    pushrod_window.add_widget(Box::new(box_widget));
+    prod.widget_store.add_widget(Box::new(box_widget));
 
     let mut box_1 = BoxWidget::new();
     box_1.set_origin(500, 50);
     box_1.set_size(200, 200);
     box_1.set_color([0.5, 0.5, 1.0, 1.0]);
     box_1.set_border([0.0, 0.0, 1.0, 1.0], 2);
-    let box_1_id = pushrod_window.add_widget(Box::new(box_1));
+    let box_1_id = prod.widget_store.add_widget(Box::new(box_1));
 
     let mut box_2 = BoxWidget::new();
     box_2.set_origin(550, 75);
     box_2.set_size(100, 50);
     box_2.set_color([0.75, 0.75, 1.0, 1.0]);
     box_2.set_border([1.0, 0.0, 1.0, 1.0], 1);
-    pushrod_window.add_widget_to_parent(Box::new(box_2), box_1_id);
+    prod.widget_store.add_widget_to_parent(Box::new(box_2), box_1_id);
 
     let mut box_3 = BoxWidget::new();
     box_3.set_origin(550, 175);
     box_3.set_size(100, 50);
     box_3.set_color([0.75, 0.75, 1.0, 1.0]);
     box_3.set_border([1.0, 0.0, 1.0, 1.0], 1);
-    pushrod_window.add_widget_to_parent(Box::new(box_3), box_1_id);
+    prod.widget_store.add_widget_to_parent(Box::new(box_3), box_1_id);
 
     let mut timer = TimerWidget::new();
     timer.set_timeout(1000);
     timer.set_enabled(true);
     timer.on_timeout(Box::new(|| eprintln!("Timer.")));
-    pushrod_window.add_widget(Box::new(timer));
+    prod.widget_store.add_widget(Box::new(timer));
 
-    prod.add_window(pushrod_window);
+//    prod.add_window(pushrod_window);
     prod.add_event_listener_for_window(Box::new(ExampleListener::new()));
 
     // Runs the main event loop