+++ src/core/main.rs
@@ -14,12 +14,13 @@
 // limitations under the License.
 
 use crate::core::point::*;
-use crate::core::window::*;
+use crate::core::widget_store::*;
 use crate::event::event::*;
 
-use opengl_graphics::GlGraphics;
+use opengl_graphics::{GlGraphics, Texture};
 use graphics::math::*;
 use piston_window::*;
+use gl::types::GLuint;
 
 use std::cell::RefCell;
 
@@ -31,9 +32,13 @@ use std::cell::RefCell;
 /// The objects contained within this structure are used by the `Pushrod` run loop, and
 /// are not intended to be modified except through methods in the `Pushrod` impl.
 pub struct Pushrod {
-    windows: RefCell<Vec<PushrodWindow>>,
+    window: PistonWindow,
+    pub widget_store: WidgetStore,
     event_listeners: RefCell<Vec<Box<EventListener>>>,
     event_list: RefCell<Vec<PushrodEvent>>,
+    texture_buf: Box<Vec<u8>>,
+    texture: Texture,
+    fbo: GLuint,
 }
 
 /// Pushrod implementation.  Create a `Pushrod::new( OpenGL )` object to create a new
@@ -68,19 +73,16 @@ pub struct Pushrod {
 /// ```
 impl Pushrod {
     /// Pushrod Object Constructor.  Takes in a single OpenGL configuration type.
-    pub fn new() -> Self {
+    pub fn new(window: PistonWindow) -> Self {
         Self {
-            windows: RefCell::new(Vec::new()),
+            window,
+            widget_store: WidgetStore::new(),
             event_listeners: RefCell::new(Vec::new()),
             event_list: RefCell::new(Vec::new()),
+            texture_buf: Box::new(vec![0u8; 1]),
+            texture: Texture::empty(&TextureSettings::new()).unwrap(),
+            fbo: 0,
         }
-    }
-
-    /// Adds a managed window to the stack.  Changes the `swap_buffers` flag to `false`, as if
-    /// this were set to `true`, the main draw loop would be extremely expensive.
-    pub fn add_window(&self, mut window: PushrodWindow) {
-        &window.prepare_buffers();
-        self.windows.borrow_mut().push(window);
     }
 
     /// Adds an event listener to the stack.  This should be an implementation of the
@@ -160,7 +162,7 @@ impl Pushrod {
     /// }
     /// ```
     pub fn add_event_listener_for_window(&self, listener: Box<EventListener>) {
-        self.event_listeners.borrow_mut().push(listener);
+//        self.event_listeners.borrow_mut().push(listener);
     }
 
     /*
@@ -174,93 +176,88 @@ impl Pushrod {
      */
 
     fn internal_handle_mouse_move(&self, point: Point) {
-        // Send the point movement to the widget event handler.
-
-        self.event_list
-            .borrow_mut()
-            .push(PushrodEvent::MouseEvent { point });
+//        // Send the point movement to the widget event handler.
+//
+//        self.event_list
+//            .borrow_mut()
+//            .push(PushrodEvent::MouseEvent { point });
     }
 
     fn internal_handle_mouse_button(&self, button: ButtonArgs) {
-        // Send the button click to the widget event handler.
-
-        if button.state == ButtonState::Press {
-            match button.button {
-                Button::Mouse(button) => {
-                    self.event_list
-                        .borrow_mut()
-                        .push(PushrodEvent::MouseDownEvent { button });
-                }
-                _ => (),
-            }
-        } else if button.state == ButtonState::Release {
-            match button.button {
-                Button::Mouse(button) => {
-                    self.event_list
-                        .borrow_mut()
-                        .push(PushrodEvent::MouseUpEvent { button });
-                }
-                _ => (),
-            }
-        }
+//        // Send the button click to the widget event handler.
+//
+//        if button.state == ButtonState::Press {
+//            match button.button {
+//                Button::Mouse(button) => {
+//                    self.event_list
+//                        .borrow_mut()
+//                        .push(PushrodEvent::MouseDownEvent { button });
+//                }
+//                _ => (),
+//            }
+//        } else if button.state == ButtonState::Release {
+//            match button.button {
+//                Button::Mouse(button) => {
+//                    self.event_list
+//                        .borrow_mut()
+//                        .push(PushrodEvent::MouseUpEvent { button });
+//                }
+//                _ => (),
+//            }
+//        }
     }
 
     fn internal_handle_mouse_scroll(&self, point: Point) {
-        // Send the mouse scroll to the widget event handler.
-
-        self.event_list
-            .borrow_mut()
-            .push(PushrodEvent::MouseScrollEvent { point });
+//        // Send the mouse scroll to the widget event handler.
+//
+//        self.event_list
+//            .borrow_mut()
+//            .push(PushrodEvent::MouseScrollEvent { point });
     }
 
     fn internal_dispatch_events(&self) {
-        for event in self.event_list.borrow_mut().iter() {
-            for listener in self.event_listeners.borrow_mut().iter() {
-                let event_mask = self.internal_derive_event_mask(event);
-
-                if listener.event_mask() & event_mask == event_mask {
-                    listener.handle_event(event);
-                }
-            }
-        }
-
-        self.event_list.borrow_mut().clear();
+//        for event in self.event_list.borrow_mut().iter() {
+//            for listener in self.event_listeners.borrow_mut().iter() {
+//                let event_mask = self.internal_derive_event_mask(event);
+//
+//                if listener.event_mask() & event_mask == event_mask {
+//                    listener.handle_event(event);
+//                }
+//            }
+//        }
+//
+//        self.event_list.borrow_mut().clear();
     }
 
     fn internal_derive_event_mask(&self, event: &PushrodEvent) -> EventMask {
-        match event {
-            PushrodEvent::MouseEvent { point: _ } => MASK_EVENT_MOUSE_MOVED,
-            PushrodEvent::MouseDownEvent { button: _ } => MASK_EVENT_MOUSE_DOWN,
-            PushrodEvent::MouseUpEvent { button: _ } => MASK_EVENT_MOUSE_UP,
-            PushrodEvent::MouseScrollEvent { point: _ } => MASK_EVENT_MOUSE_SCROLL,
-        }
+//        match event {
+//            PushrodEvent::MouseEvent { point: _ } => MASK_EVENT_MOUSE_MOVED,
+//            PushrodEvent::MouseDownEvent { button: _ } => MASK_EVENT_MOUSE_DOWN,
+//            PushrodEvent::MouseUpEvent { button: _ } => MASK_EVENT_MOUSE_UP,
+//            PushrodEvent::MouseScrollEvent { point: _ } => MASK_EVENT_MOUSE_SCROLL,
+//        }
+        0
     }
 
-    fn recursive_draw(
-        &self,
-        pushrod_window: &mut PushrodWindow,
-        widget_id: i32,
-        context: Context,
-        graphics: &mut GlGraphics,
-    ) {
-        let parents_of_widget = pushrod_window.get_children_of(widget_id);
+    fn handle_resize(&mut self, width: u32, height: u32) {
+        self.texture_buf = Box::new(vec![0u8; width as usize * height as usize]);
+        self.texture = Texture::from_memory_alpha(&self.texture_buf, width, height,
+        &TextureSettings::new()).unwrap();
 
-        if parents_of_widget.len() == 0 {
-            return;
-        }
+        // I hate this code.  However, this does prepare a texture so that an image can be
+        // drawn on it.  Since it's in memory, it means that the texture only gets recreated once
+        // per resize.
+        unsafe {
+            let mut fbos: [GLuint; 1] = [0];
 
-        for pos in 0..parents_of_widget.len() {
-            let paint_id = parents_of_widget[pos];
-            let paint_widget = &mut pushrod_window.widgets[paint_id as usize];
+            gl::GenFramebuffers(1, fbos.as_mut_ptr());
+            self.fbo = fbos[0];
 
-            if &paint_widget.widget.is_invalidated() == &true {
-                eprintln!("[Invalidated] Painting {}", paint_id);
-                &paint_widget.widget.draw(context, graphics);
-            }
-
-            if parents_of_widget[pos] != widget_id {
-                self.recursive_draw(pushrod_window, paint_id, context, graphics);
-            }
+            gl::BindFramebuffer(gl::FRAMEBUFFER, self.fbo);
+            gl::FramebufferTexture2D(gl::FRAMEBUFFER,
+                gl::COLOR_ATTACHMENT0, gl::TEXTURE_2D,
+                self.texture.get_id(),
+                0);
         }
     }
 
@@ -282,106 +279,93 @@ impl Pushrod {
     /// This event is handled window-by-window.  Once a window has processed all of its pending
     /// events, the next window is then processed.  No particular window takes precidence - any
     /// window that has events to process gets handled in order.
-    pub fn run(&self) {
+    pub fn run(&mut self) {
         let mut gl: GlGraphics = GlGraphics::new(OpenGL::V3_2);
         let mut last_widget_id = -1;
         let mut previous_mouse_position: Point = make_origin_point();
 
-        for (_window_id, pushrod_window) in self.windows.borrow_mut().iter_mut().enumerate() {
-            while let Some(event) = &pushrod_window.window.next() {
+//        for (_window_id, pushrod_window) in self.windows.borrow_mut().iter_mut().enumerate() {
+        while let Some(event) = self.window.next() {
                 event.mouse_cursor(|x, y| {
-                    let mouse_point = make_point_f64(x, y);
-
-                    if mouse_point.x != previous_mouse_position.x
-                        || mouse_point.y != previous_mouse_position.y
-                    {
-                        previous_mouse_position = mouse_point.clone();
-
-                        self.internal_handle_mouse_move(mouse_point.clone());
-
-                        let current_widget_id = pushrod_window.get_widget_id_for_point(mouse_point);
-                        let current_parent_for_widget =
-                            pushrod_window.get_parent_of(current_widget_id);
-
-                        if current_widget_id != last_widget_id {
-                            if last_widget_id != -1 {
-                                pushrod_window.mouse_exited_for_id(last_widget_id);
-                            }
-
-                            last_widget_id = current_widget_id;
-
-                            if last_widget_id != -1 {
-                                pushrod_window.mouse_entered_for_id(last_widget_id);
-                            }
-                        }
-
-                        eprintln!(
-                            "Widget IDs: current={} parent={} children={:?}",
-                            current_widget_id,
-                            current_parent_for_widget,
-                            pushrod_window.get_children_of(current_widget_id)
-                        );
-                    }
+//                    let mouse_point = make_point_f64(x, y);
+//
+//                    if mouse_point.x != previous_mouse_position.x
+//                        || mouse_point.y != previous_mouse_position.y
+//                    {
+//                        previous_mouse_position = mouse_point.clone();
+//
+//                        self.internal_handle_mouse_move(mouse_point.clone());
+//
+//                        let current_widget_id = pushrod_window.get_widget_id_for_point(mouse_point);
+//                        let current_parent_for_widget =
+//                            pushrod_window.get_parent_of(current_widget_id);
+//
+//                        if current_widget_id != last_widget_id {
+//                            if last_widget_id != -1 {
+//                                pushrod_window.mouse_exited_for_id(last_widget_id);
+//                            }
+//
+//                            last_widget_id = current_widget_id;
+//
+//                            if last_widget_id != -1 {
+//                                pushrod_window.mouse_entered_for_id(last_widget_id);
+//                            }
+//                        }
+//
+//                        eprintln!(
+//                            "Widget IDs: current={} parent={} children={:?}",
+//                            current_widget_id,
+//                            current_parent_for_widget,
+//                            pushrod_window.get_children_of(current_widget_id)
+//                        );
+//                    }
                 });
 
                 event.button(|button| {
-                    self.internal_handle_mouse_button(button);
+//                    self.internal_handle_mouse_button(button);
                 });
 
                 event.mouse_scroll(|x, y| {
-                    let mouse_point = make_point_f64(x, y);
-
-                    self.internal_handle_mouse_scroll(mouse_point.clone());
-
-                    if last_widget_id != -1 {
-                        pushrod_window.mouse_scrolled_for_id(last_widget_id, mouse_point.clone());
-                    }
+//                    let mouse_point = make_point_f64(x, y);
+//
+//                    self.internal_handle_mouse_scroll(mouse_point.clone());
+//
+//                    if last_widget_id != -1 {
+//                        pushrod_window.mouse_scrolled_for_id(last_widget_id, mouse_point.clone());
+//                    }
                 });
 
-                event.resize(|width, height| {
-                    pushrod_window.handle_resize(width as u32, height as u32);
-                    pushrod_window.invalidate_all_widgets();
-                });
+            event.resize(|width, height| {
+                self.handle_resize(width as u32, height as u32);
+                self.widget_store.invalidate_all_widgets();
+            });
 
-                // Dispatch events here in the bus
-                self.internal_dispatch_events();
+            // Dispatch events here in the bus
+            self.internal_dispatch_events();
 
-                // FPS loop handling
+            // FPS loop handling
 
-                event.render(|args| {
-                    let is_invalidated = pushrod_window
-                        .widgets
-                        .iter_mut()
-                        .map(|x| x.widget.is_invalidated())
-                        .find(|x| x == &true)
-                        .unwrap_or(false);
-
-                    if is_invalidated {
-                        // Swap frame buffers so that we're drawing to the currently active
-                        // pushrod window's texture.
-                        unsafe {
-                            gl::BindFramebuffer(gl::FRAMEBUFFER, pushrod_window.fbo);
-                        }
-
-                        // Draw the graphics onto the texture, this is _super_ fast.
-                        gl.draw(args.viewport(), |c, g| {
-                            self.recursive_draw(pushrod_window, 0, c, g);
-                        });
-                    }
-
-                    // Switch back to the main screen's window framebuffer.
+            event.render(|args| {
+                if self.widget_store.needs_repaint() {
                     unsafe {
-                        gl::BindFramebuffer(gl::FRAMEBUFFER, 0);
+                        gl::BindFramebuffer(gl::FRAMEBUFFER, self.fbo);
                     }
 
-                    // Continuously draw the screen's texture as an image back to the main screen.
                     gl.draw(args.viewport(), |c, g| {
-                        clear([1.0, 1.0, 1.0, 0.0], g);
-                        let flipped = c.transform.prepend_transform(scale(1., -1.));
-                        Image::new().draw(&pushrod_window.texture, &c.draw_state, flipped, g);
+                        self.widget_store.draw(0, c, g);
                     });
+                }
+
+                unsafe {
+                    gl::BindFramebuffer(gl::FRAMEBUFFER, 0);
+                }
+
+                gl.draw(args.viewport(), |c, g| {
+                    clear([1.0, 1.0, 1.0, 0.0], g);
+                    let flipped = c.transform.prepend_transform(scale(1.0, -1.0));
+                    Image::new().draw(&self.texture, &c.draw_state, flipped, g);
                 });
-            }
+            });
         }
     }
 }